Implemented ToggleDPS
The designers of the STC have been really clever: to access DPTR1,
there's a bit you have to toggle. The problem is, it's not in a
bit-addressable register. But they've made it bit 0 - and defined bit 1
as 0 all the time. So you can INC rAUXR1 to toggle which DPTR you're
accessing!